pyproject.toml
[project]
name = "mcp-server-fastmcp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi-mcp>=0.4.0",
    "fastapi[standard]>=0.121.2",
    "fastmcp>=2.13.0.2",
    "google-genai>=1.50.1",
    "openai>=2.8.0",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
]
